chore: remove blog data from providers, cache snippets
